Refactor: 퍼센트 표현방식을 실수에서 정수로 변경

+++ src/main/java/com/example/savingsalt/goal/domain/dto/GoalCertificationStatisticsResDto.java
@@ -25,7 +25,7 @@ public class GoalCertificationStatisticsResDto {
     private Set<String> monthlyCertifications;
 
     // 월간 인증 내용 백분율
-    private Map<String, Double> monthlyCertificationPercentages;
+    private Map<String, Long> monthlyCertificationPercentages;
 
     // 일일 인증 금액
     private Long dailyAmount;
